+++ pkg/pull/model.go
@@ -19,8 +19,9 @@ package pull
 
 import (
 	"encoding/json"
+	"time"
 
-	over "github.com/Trendyol/overlog"
+	over "github.com/adrienaury/zeromdc"
 	"github.com/cgi-fr/jsonline/pkg/jsonline"
 	"github.com/rs/zerolog/log"
 )
@@ -110,6 +111,7 @@ func (er ExportedRow) GetOrNil(key string) interface{} {
 type ExecutionStats interface {
 	GetLinesPerStepCount() map[string]int
 	GetFiltersCount() int
+	GetDuration() time.Duration
 
 	ToJSON() []byte
 }
@@ -117,6 +119,7 @@ type ExecutionStats interface {
 type stats struct {
 	LinesPerStepCount map[string]int `json:"linesPerStepCount"`
 	FiltersCount      int            `json:"filtersCount"`
+	Duration          time.Duration  `json:"duration"`
 }
 
 func (s *stats) ToJSON() []byte {
@@ -135,6 +138,10 @@ func (s *stats) GetFiltersCount() int {
 	return s.FiltersCount
 }
 
+func (s *stats) GetDuration() time.Duration {
+	return s.Duration
+}
+
 func IncLinesPerStepCount(step string) {
 	stats := getStats()
 	stats.LinesPerStepCount[step]++
@@ -143,6 +150,11 @@ func IncLinesPerStepCount(step string) {
 func IncFiltersCount() {
 	stats := getStats()
 	stats.FiltersCount++
+}
+
+func SetDuration(duration time.Duration) {
+	stats := getStats()
+	stats.Duration = duration
 }
 
 func getStats() *stats {
@@ -154,6 +166,7 @@ func getStats() *stats {
 	return &stats{
 		LinesPerStepCount: map[string]int{},
 		FiltersCount:      0,
+		Duration:          0,
 	}
 }
 
@@ -168,12 +181,4 @@ func Compute() ExecutionStats {
 
 func Reset() {
 	over.MDC().Set("stats", &stats{FiltersCount: 0, LinesPerStepCount: map[string]int{}})
-}
-
-func MutualizeStats(s stats) {
-	stats := getStats()
-	stats.FiltersCount += s.FiltersCount
-	for key, value := range s.LinesPerStepCount {
-		stats.LinesPerStepCount[key] += value
-	}
 }